refactor(fan): extract shared fan-control helpers into domain/fan_control
Move compute_fan_speed, evaluate_temp_override, compute_wind_offset,
FAN_VPD_STAGE_DEFAULTS, and the stage-aware VPD target resolution out of
circulation_fan_coordinator into a shared domain.fan_control module so the
forthcoming ExhaustFanController can reuse one implementation.

The circulation coordinator re-exports the helpers, so existing importers and
unit tests keep working unchanged; only the determine_coordinator_stage patch
target moved with the code. No behavioral change to circulation fan control.

Closes #473
