+++ custom_components/growspace_manager/circulation_fan_coordinator.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 from collections.abc import Callable
 from datetime import timedelta
 import logging
-import math
 import time
 from typing import TYPE_CHECKING
 
@@ -15,9 +14,15 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.event import async_track_time_interval
 
-from .const import FanRegulationMode, PlantStage
+from .const import FanRegulationMode
 from .domain.day_night import DayNightTracker
-from .domain.stage_calculator import determine_coordinator_stage
+from .domain.fan_control import (
+    FAN_VPD_STAGE_DEFAULTS,
+    compute_fan_speed,
+    compute_wind_offset,
+    evaluate_temp_override,
+    resolve_stage_vpd_target,
+)
 
 if TYPE_CHECKING:
     from .coordinator import GrowspaceCoordinator
@@ -27,88 +32,14 @@ _LOGGER = logging.getLogger(__name__)
 
 _TICK_INTERVAL = timedelta(seconds=10)
 
-# Per-stage VPD targets (day / night) for dynamic VPD mode.
-# Values are midpoints of the tightest Bayesian optimal range for each stage.
-FAN_VPD_STAGE_DEFAULTS: dict[PlantStage, dict[str, float]] = {
-    PlantStage.SEEDLING:      {"day": 0.60, "night": 0.60},
-    PlantStage.CLONE:         {"day": 0.50, "night": 0.50},
-    PlantStage.MOTHER:        {"day": 0.70, "night": 0.60},
-    PlantStage.VEG:           {"day": 0.70, "night": 0.60},
-    PlantStage.FLOWER_EARLY:  {"day": 1.15, "night": 1.00},
-    PlantStage.FLOWER_MID:    {"day": 1.20, "night": 1.00},
-    PlantStage.FLOWER_LATE:   {"day": 1.25, "night": 1.05},
-    PlantStage.DRY:           {"day": 0.95, "night": 0.95},
-    PlantStage.CURE:          {"day": 0.75, "night": 0.75},
-}
-
-
-def evaluate_temp_override(
-    current_temp: float,
-    critical_temp_low: float | None,
-    critical_temp_high: float | None,
-    hysteresis: float,
-    override_active: bool,
-    override_direction: str | None,
-    vpd_speed: int,
-    min_speed: int,
-    max_speed: int,
-) -> tuple[int, bool, str | None]:
-    """Apply temperature safety override logic for VPD mode.
-
-    Returns (final_speed, new_override_active, new_override_direction).
-    Override direction is "high" or "low" when active, None otherwise.
-    """
-    if critical_temp_low is None and critical_temp_high is None:
-        return vpd_speed, False, None
-
-    if not override_active:
-        if critical_temp_high is not None and current_temp > critical_temp_high:
-            return max_speed, True, "high"
-        if critical_temp_low is not None and current_temp < critical_temp_low:
-            return min_speed, True, "low"
-        return vpd_speed, False, None
-
-    if override_direction == "high":
-        if critical_temp_high is not None and current_temp <= critical_temp_high - hysteresis:
-            return vpd_speed, False, None
-        return max_speed, True, "high"
-
-    # override_direction == "low"
-    if critical_temp_low is not None and current_temp >= critical_temp_low + hysteresis:
-        return vpd_speed, False, None
-    return min_speed, True, "low"
-
-
-def compute_wind_offset(
-    amplitude_pct: int,
-    elapsed_seconds: float,
-    period_seconds: int,
-) -> float:
-    """Compute wind offset as amplitude × sin(2π × elapsed / period)."""
-    return amplitude_pct * math.sin(2 * math.pi * elapsed_seconds / period_seconds)
-
-
-def compute_fan_speed(
-    value: float,
-    target: float,
-    tolerance: float,
-    min_speed: int,
-    max_speed: int,
-) -> int:
-    """Compute fan speed via linear mapping of value relative to target band.
-
-    Below (target - tolerance): min_speed
-    Above (target + tolerance): max_speed
-    Inside the band: linearly interpolated
-    """
-    lower = target - tolerance
-    upper = target + tolerance
-    if value <= lower:
-        return min_speed
-    if value >= upper:
-        return max_speed
-    t = (value - lower) / (upper - lower)
-    return round(min_speed + t * (max_speed - min_speed))
+# Re-exported from .domain.fan_control so existing importers keep working.
+__all__ = [
+    "FAN_VPD_STAGE_DEFAULTS",
+    "CirculationFanCoordinator",
+    "compute_fan_speed",
+    "compute_wind_offset",
+    "evaluate_temp_override",
+]
 
 
 class CirculationFanCoordinator:
@@ -260,18 +191,9 @@ class CirculationFanCoordinator:
         plants = self.main_coordinator.services.growspaces.get_growspace_plants(
             self.growspace_id
         )
-        if not plants:
-            return cfg.vpd_target
-
-        stage = determine_coordinator_stage(plants)
-        day_key = "day" if is_day else "night"
-        override = cfg.stage_vpd_overrides.get(stage.value)
-        if override:
-            return override[day_key]
-        stage_entry = FAN_VPD_STAGE_DEFAULTS.get(stage)
-        if stage_entry:
-            return stage_entry[day_key]
-        return cfg.vpd_target
+        return resolve_stage_vpd_target(
+            plants, cfg.stage_vpd_overrides, cfg.vpd_target, is_day
+        )
 
     def _read_sensor(self, mode: FanRegulationMode) -> float | None:
         """Read the first available sensor value for the given regulation mode."""
